rol y rol usuario

+++ examenAsoWebJPA/ejbModule/py/edu/ucsa/ejb/dto/RolDTO.java
@@ -2,12 +2,14 @@ package py.edu.ucsa.ejb.dto;
 
 public class RolDTO {
 
-	private int id;
+	private Long id;
 	private String nombreRol;
-	public int getId() {
+
+		
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getNombreRol() {
